refactor: use uint64's instead of uint32's

+++ id.go
@@ -6,17 +6,17 @@ type ClientID = uint64
 
 type ID struct {
 	client ClientID
-	clock  uint32
+	clock  uint64
 }
 
-func newID(client ClientID, clock uint32) *ID {
+func newID(client ClientID, clock uint64) *ID {
 	return &ID{
 		client: client,
 		clock:  clock,
 	}
 }
 
-func NewID(client ClientID, clock uint32) *ID {
+func NewID(client ClientID, clock uint64) *ID {
 	return newID(client, clock)
 }
